// openshift/tekton-analyser/src/types.ts
export interface Issue {
  severity: 'error' | 'warning';
  code: string;
  message: string;
  kind?: string;
  name?: string;
  namespace?: string;
  source?: string;
  path?: string;
}
// ...
export interface OtherResource {
  apiVersion: string;
  kind: string;
  name?: string;
  namespace?: string;
  source?: string;
  tekton: boolean;
}